refactore structure of project and add google-formatter-plugin to sub-projects
