--- common/src/main/java/com/sampledomain/common/utils/CommonUtil.java
@@ -1,25 +1,0 @@
-package com.sampledomain.common.utils;
-
-import java.util.Collections;
-import java.util.List;
-import java.util.Optional;
-
-public final class CommonUtil {
-  private CommonUtil() {}
-
-  public static <T> List<T> listOrEmpty(List<T> list) {
-    return Optional.ofNullable(list).orElse(Collections.emptyList());
-  }
-
-  public static boolean nullOrEmpty(String string) {
-    return string == null || string.isEmpty();
-  }
-
-  public static boolean nullOrEmpty(List<?> list) {
-    return list == null || list.isEmpty();
-  }
-
-  public static boolean nullOrEmpty(Object object) {
-    return object == null || nullOrEmpty(object.toString());
-  }
-}
