+++ url-counter-service/src/main/java/com/echitgar/URLCounterServiceApplication.java
@@ -1,0 +1,11 @@
+package com.echitgar;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class URLCounterServiceApplication {
+  public static void main(String[] args) {
+    SpringApplication.run(URLCounterServiceApplication.class, args);
+  }
+}
