+++ url-counter-service/src/main/java/com/echitgar/kafka/config/KafkaTopicConfig.java
@@ -1,0 +1,18 @@
+package com.echitgar.kafka.config;
+
+import org.apache.kafka.clients.admin.NewTopic;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.kafka.config.TopicBuilder;
+
+@Configuration
+public class KafkaTopicConfig {
+  @Value(value = "${message.topic1}")
+  private String topic1;
+
+  @Bean
+  public NewTopic topic1() {
+    return TopicBuilder.name(topic1).build();
+  }
+}
