--- url-counter-service/src/main/java/com/sampledomain/URLCounterServiceApplication.java
@@ -1,11 +1,0 @@
-package com.sampledomain;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class URLCounterServiceApplication {
-  public static void main(String[] args) {
-    SpringApplication.run(URLCounterServiceApplication.class, args);
-  }
-}
