--- url-counter-service/src/main/java/com/sampledomain/kafka/config/KafkaConsumerConfig.java
@@ -1,51 +1,0 @@
-package com.sampledomain.kafka.config;
-
-import com.sampledomain.messages.Message;
-import java.util.HashMap;
-import java.util.Map;
-import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.common.serialization.StringDeserializer;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
-import org.springframework.kafka.core.ConsumerFactory;
-import org.springframework.kafka.core.DefaultKafkaConsumerFactory;
-import org.springframework.kafka.support.converter.RecordMessageConverter;
-import org.springframework.kafka.support.converter.StringJsonMessageConverter;
-import org.springframework.kafka.support.serializer.ErrorHandlingDeserializer;
-import org.springframework.kafka.support.serializer.JsonDeserializer;
-
-@Configuration
-public class KafkaConsumerConfig {
-  @Value("${spring.kafka.bootstrap-servers}")
-  private String bootstrapServers;
-
-  @Bean
-  public RecordMessageConverter converter() {
-    StringJsonMessageConverter converter = new StringJsonMessageConverter();
-    return converter;
-  }
-
-  public Map<String, Object> consumerConfig() {
-    Map<String, Object> props = new HashMap<>();
-    props.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers);
-    props.put(ErrorHandlingDeserializer.KEY_DESERIALIZER_CLASS, StringDeserializer.class);
-    props.put(ErrorHandlingDeserializer.VALUE_DESERIALIZER_CLASS, JsonDeserializer.class);
-    return props;
-  }
-
-  @Bean
-  public ConsumerFactory<String, Message> consumerFactory() {
-    return new DefaultKafkaConsumerFactory<>(
-        consumerConfig(), new StringDeserializer(), new JsonDeserializer<>(Message.class));
-  }
-
-  @Bean
-  public ConcurrentKafkaListenerContainerFactory<String, Message> kafkaListenerContainerFactory() {
-    ConcurrentKafkaListenerContainerFactory<String, Message> factory =
-        new ConcurrentKafkaListenerContainerFactory<>();
-    factory.setConsumerFactory(consumerFactory());
-    return factory;
-  }
-}
