--- url-counter-service/src/main/java/com/sampledomain/log/LogManager.java
@@ -1,8 +1,0 @@
-package com.sampledomain.log;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class LogManager {
-  public static final Logger LOG = LoggerFactory.getLogger(LogManager.class);
-}
