+++ url-counter-service/src/test/java/com/echitgar/URLCounterServiceApplicationTests.java
@@ -1,0 +1,6 @@
+package com.echitgar;
+
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class URLCounterServiceApplicationTests {}
