--- url-counter-service/src/test/java/com/sampledomain/URLCounterServiceApplicationTests.java
@@ -1,6 +1,0 @@
-package com.sampledomain;
-
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class URLCounterServiceApplicationTests {}
