+++ url-generator-service/pom.xml
@@ -4,19 +4,28 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>com.sampledomain</groupId>
-        <artifactId>UrlAnalytics</artifactId>
-        <version>0.0.1</version>
+        <groupId>com.echitgar</groupId>
+        <artifactId>url-analytics</artifactId>
+        <version>0.0.2</version>
     </parent>
 
-    <artifactId>URLGenerator</artifactId>
-    <name>URLGenerator</name>
+    <artifactId>url-generator-service</artifactId>
+    <name>url-generator-service</name>
 
     <dependencies>
         <dependency>
-            <groupId>com.sampledomain</groupId>
+            <groupId>com.echitgar</groupId>
             <artifactId>common</artifactId>
-            <version>0.0.1</version>
+            <version>0.0.2</version>
         </dependency>
     </dependencies>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>com.theoryinpractise</groupId>
+                <artifactId>googleformatter-maven-plugin</artifactId>
+            </plugin>
+        </plugins>
+    </build>
 </project>
