+++ url-generator-service/src/main/java/com/echitgar/bootstrap/BootstrapService.java
@@ -1,0 +1,17 @@
+package com.echitgar.bootstrap;
+
+import com.echitgar.send.UrlSendManager;
+import org.springframework.stereotype.Service;
+
+@Service
+public class BootstrapService {
+  private final UrlSendManager urlSendManager;
+
+  public BootstrapService(UrlSendManager urlSendManager) {
+    this.urlSendManager = urlSendManager;
+  }
+
+  public void run() {
+    urlSendManager.send();
+  }
+}
