+++ url-generator-service/src/main/java/com/echitgar/log/LogManager.java
@@ -1,0 +1,8 @@
+package com.echitgar.log;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class LogManager {
+  public static final Logger LOG = LoggerFactory.getLogger(LogManager.class);
+}
