--- url-generator-service/src/main/java/com/sampledomain/bootstrap/BootstrapService.java
@@ -1,17 +1,0 @@
-package com.sampledomain.bootstrap;
-
-import com.sampledomain.send.UrlSendManager;
-import org.springframework.stereotype.Service;
-
-@Service
-public class BootstrapService {
-  private final UrlSendManager urlSendManager;
-
-  public BootstrapService(UrlSendManager urlSendManager) {
-    this.urlSendManager = urlSendManager;
-  }
-
-  public void run() {
-    urlSendManager.send();
-  }
-}
