--- url-generator-service/src/main/java/com/sampledomain/kafka/config/KafkaProducerConfig.java
@@ -1,38 +1,0 @@
-package com.sampledomain.kafka.config;
-
-import java.util.HashMap;
-import java.util.Map;
-import org.apache.kafka.clients.producer.ProducerConfig;
-import org.apache.kafka.common.serialization.StringSerializer;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.core.DefaultKafkaProducerFactory;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.kafka.core.ProducerFactory;
-import org.springframework.kafka.support.serializer.JsonSerializer;
-
-@Configuration
-public class KafkaProducerConfig {
-  @Value("${spring.kafka.bootstrap-servers}")
-  private String bootstrapServers;
-
-  public Map<String, Object> producerConfig() {
-    Map<String, Object> props = new HashMap<>();
-    props.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers);
-    props.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
-    props.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, JsonSerializer.class);
-    return props;
-  }
-
-  @Bean
-  public ProducerFactory<String, Object> producerFactory() {
-    return new DefaultKafkaProducerFactory<>(producerConfig());
-  }
-
-  @Bean
-  public KafkaTemplate<String, Object> kafkaTemplate(
-      ProducerFactory<String, Object> producerFactory) {
-    return new KafkaTemplate<>(producerFactory);
-  }
-}
