+++ url-generator-service/src/test/java/com/echitgar/URLGeneratorServiceApplicationTests.java
@@ -1,0 +1,6 @@
+package com.echitgar;
+
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class URLGeneratorServiceApplicationTests {}
