--- url-generator-service/src/test/java/com/sampledomain/URLGeneratorServiceApplicationTests.java
@@ -1,6 +1,0 @@
-package com.sampledomain;
-
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class URLGeneratorServiceApplicationTests {}
